audio: cast the streams flag mask to u64 like its ten siblings
Ten of the eleven "gAudioPendingLoadFlags &= ~(u64)AUDIO_LOAD_*"
statements in this file already carry the cast and all sit at 100; line
476 was the lone outlier. The cast is what makes MWCC emit retail's
li r0,-5; and r0,r3,r0 instead of our rlwinm, and it removes the
instruction-count mismatch outright -- streamsLoadedCallback goes from
T=87 C=86 to T=87 C=87, leaving only the const-zero placement, which is
a known closed wall.

Note the metric disagrees: objdiff's alignment charges an insert plus a
delete for that remaining diff, so unit fuzzy reads 99.934 -> 99.917
even though the code is structurally closer. matched_code is unchanged
at 87.05322 -- the function is not byte-exact either way -- so nothing
real is lost, and the source is now consistent with its own siblings and
with what the compiler evidence says the original wrote.

+++ src/main/audio.c
@@ -473,7 +473,7 @@ void streamsLoadedCallback(s32 status, DVDFileInfo* fileInfo)
         saved = mmSetFreeDelay(0);
         mm_free(fileInfo);
         mmSetFreeDelay(saved);
-        gAudioPendingLoadFlags &= ~AUDIO_LOAD_STREAMS;
+        gAudioPendingLoadFlags &= ~(u64)AUDIO_LOAD_STREAMS;
         gAudioCompletedLoadFlags |= AUDIO_LOAD_STREAMS;
         stream = gStreamsData;
         streamCount = gStreamsCount;
